src/pyproject.toml
[tool.poetry]
name = "spritebot"
version = "0.1.0"
description = ""
authors = ["Joanie <greystorm101@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
